+++ fbw-a380x/src/systems/instruments/src/MFD/pages/FMS/MfdFmsFuelLoad.tsx
@@ -66,8 +66,6 @@ export class MfdFmsFuelLoad extends FmsPage<MfdFmsFuelLoadProps> {
 
   private readonly destEfob = Subject.create<string>('---.-');
 
-  private readonly destEfobBelowMin = Subject.create(false);
-
   private readonly altnIcao = Subject.create<string>('----');
 
   private readonly altnEta = Subject.create<string>('--:--');
@@ -96,14 +94,6 @@ export class MfdFmsFuelLoad extends FmsPage<MfdFmsFuelLoadProps> {
       this.costIndex.set(this.loadedFlightPlan.performanceData.costIndex);
     }
 
-    if (!this.props.fmcService.master.fmgc.data.finalFuelIsPilotEntered.get()) {
-      // Calculate final res fuel for 00:30 time
-      this.props.fmcService.master.fmgc.data.finalFuelWeightCalculated.set(4_650); // FIXME
-    }
-
-    // FIXME calculate altn fuel
-    this.props.fmcService.master.fmgc.data.alternateFuelCalculated.set(6_500); // FIXME
-
     this.updateDestAndAltnPredictions();
   }
 
@@ -122,9 +112,6 @@ export class MfdFmsFuelLoad extends FmsPage<MfdFmsFuelLoadProps> {
       );
       const destEfob = this.props.fmcService.master.fmgc.getDestEFOB(true);
       this.destEfob.set(destEfob !== null ? destEfob.toFixed(1) : '---.-');
-      this.destEfobBelowMin.set(
-        destEfob * 1_000 < (this.props.fmcService.master.fmgc.data.minimumFuelAtDestination.get() ?? 0),
-      );
     }
 
     if (this.loadedFlightPlan.alternateDestinationAirport) {
@@ -487,7 +474,9 @@ export class MfdFmsFuelLoad extends FmsPage<MfdFmsFuelLoadProps> {
                   <div class="mfd-label bigger green mfd-fms-fuel-load-dest-grid-middle-cell">{this.destIcao}</div>
                   <div class="mfd-label bigger green mfd-fms-fuel-load-dest-grid-middle-cell">{this.destEta}</div>
                   <div class="mfd-label-value-container mfd-fms-fuel-load-dest-grid-efob-cell">
-                    <span class={{ 'mfd-value': true, amber: this.destEfobBelowMin }}>{this.destEfob}</span>
+                    <span class={{ 'mfd-value': true, amber: this.props.fmcService.master.fmgc.data.destEfobBelowMin }}>
+                      {this.destEfob}
+                    </span>
                     <span class="mfd-label-unit mfd-unit-trailing">T</span>
                   </div>
                   <div class="mfd-label" style="text-align: center; align-self: center;">