+++ fbw-a380x/src/systems/instruments/src/MFD/pages/FMS/MfdFmsPerf.tsx
@@ -404,7 +404,7 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
 
   private destEta = Subject.create<string>('--:--');
 
-  private destEfob = Subject.create<string>('--.-');
+  private destEfob = Subject.create<string>('---.-');
 
   private readonly crzPreSelManagedGreenLine1 = MappedSubject.create(
     ([fp, pSpeed, pMach]) =>
@@ -1019,7 +1019,7 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
             );
           }
           this.destEta.set(destEta);
-          this.destEfob.set(this.props.fmcService.master?.fmgc.getDestEFOB(true).toFixed(1) ?? '---.-');
+          this.destEfob.set(this.props.fmcService.master?.fmgc.getDestEFOB(true)?.toFixed(1) ?? '---.-');
 
           // Update DES speed table
           if (this.activeFlightPhase.get() < FmgcFlightPhase.Descent) {
@@ -2288,7 +2288,14 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                   <span class="mfd-label green bigger">{this.destAirportIdent}</span>
                   <span class="mfd-label green bigger">{this.destEta}</span>
                   <div class="mfd-label-value-container">
-                    <span class="mfd-value">{this.destEfob}</span>
+                    <span
+                      class={{
+                        'mfd-value': true,
+                        amber: this.props.fmcService.master.fmgc.data.destEfobBelowMin,
+                      }}
+                    >
+                      {this.destEfob}
+                    </span>
                     <span class="mfd-label-unit mfd-unit-trailing">T</span>
                   </div>
                   <div style="display: flex; flex-direction: row;">
@@ -2470,7 +2477,14 @@ export class MfdFmsPerf extends FmsPage<MfdFmsPerfProps> {
                   <span class="mfd-label green bigger">{this.destAirportIdent}</span>
                   <span class="mfd-label green bigger">{this.destEta}</span>
                   <div class="mfd-label-value-container">
-                    <span class="mfd-value">{this.destEfob}</span>
+                    <span
+                      class={{
+                        'mfd-value': true,
+                        amber: this.props.fmcService.master.fmgc.data.destEfobBelowMin,
+                      }}
+                    >
+                      {this.destEfob}
+                    </span>
                     <span class="mfd-label-unit mfd-unit-trailing">T</span>
                   </div>
                   <div style="display: flex; flex-direction: row;">